# Fix race in gc sweep