Removes extra goroutine and calls removal and scan in same thread