@dmcgowan 2017-10-30
Rebased on master with race fixes

GC benchmarks with this change
```
benchmark                                old ns/op     new ns/op     delta
BenchmarkGarbageCollect/10-Sets-8        761321        525422        -30.99%
BenchmarkGarbageCollect/100-Sets-8       6642323       5667259       -14.68%
BenchmarkGarbageCollect/1000-Sets-8      69455007      57987159      -16.51%
BenchmarkGarbageCollect/10000-Sets-8     818302653     733116817     -10.41%

benchmark                                old allocs     new allocs     delta
BenchmarkGarbageCollect/10-Sets-8        2485           2474           -0.44%
BenchmarkGarbageCollect/100-Sets-8       24025          24004          -0.09%
BenchmarkGarbageCollect/1000-Sets-8      258620         258589         -0.01%
BenchmarkGarbageCollect/10000-Sets-8     2805237        2805273        +0.00%

benchmark                                old bytes     new bytes     delta
BenchmarkGarbageCollect/10-Sets-8        144948        133868        -7.64%
BenchmarkGarbageCollect/100-Sets-8       1320973       1228000       -7.04%
BenchmarkGarbageCollect/1000-Sets-8      16509058      14964996      -9.35%
BenchmarkGarbageCollect/10000-Sets-8     174331384     157471128     -9.67%
```